+++ .idea/misc.xml
@@ -3,21 +3,19 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Users/jun/AndroidStudioProjects/SomethingNew-Android/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.2155" />
         <entry key="..\:/android_practice/UnithonProject/app/src/main/res/layout/activity_login.xml" value="0.25" />
         <entry key="..\:/android_practice/UnithonProject/app/src/main/res/layout/content_login.xml" value="0.18568840579710144" />
         <entry key="..\:/temp_project/SomethingNew-Android/app/src/main/res/layout/activity_address_api_web_view.xml" value="0.36770833333333336" />
         <entry key="..\:/temp_project/SomethingNew-Android/app/src/main/res/layout/activity_call.xml" value="0.20625" />
         <entry key="..\:/temp_project/SomethingNew-Android/app/src/main/res/layout/activity_location.xml" value="0.33" />
         <entry key="..\:/temp_project/SomethingNew-Android/app/src/main/res/layout/activity_login.xml" value="0.22" />
-        <entry key="..\:/temp_project/SomethingNew-Android/app/src/main/res/layout/activity_main.xml" value="0.16614583333333333" />
+        <entry key="..\:/temp_project/SomethingNew-Android/app/src/main/res/layout/activity_main.xml" value="0.33" />
+        <entry key="..\:/temp_project/SomethingNew-Android/app/src/main/res/layout/fragment_map.xml" value="0.18802083333333333" />
         <entry key="..\:/temp_project/SomethingNew-Android/app/src/main/res/layout/item_location.xml" value="0.36770833333333336" />
       </map>
     </option>
   </component>
-
   <component name="ProjectRootManager" version="2" languageLevel="JDK_11" project-jdk-name="1.8" project-jdk-type="JavaSDK">
-
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
