+++ app/src/main/java/com/unithon/somethingnew/presentation/call/CallActivity.kt
@@ -30,45 +30,6 @@ class CallActivity(override val layoutResId: Int = R.layout.activity_call) :
         setTheme(R.style.FullScreen)
         setStatusBarTransparent(this, binding.rootView)
 
-        val permissionListener: PermissionListener = object : PermissionListener {
-            override fun onPermissionGranted() {
-                // 접근허용 시 실행할 코드
-                with(binding) {
-                    remonCall = RemonCall.builder()
-                        .context(this@CallActivity)
-                        .localView(localView)        //나의 Video Renderer
-                        .remoteView(remoteView)      //상대방 video Renderer
-                        .videoCodec("VP8")
-                        .videoWidth(640)
-                        .videoHeight(480)
-                        .serviceId("SERVICEID1")    // RemoteMonster 사이트에서 등록했던 당신의 id를 입력하세요.
-                        .key("1234567890")    // RemoteMonster로부터 받은 당신의 key를 입력하세요.
-                        .build()
-                    val channelId = "unithon"//intent.getStringExtra("channelId")
-
-                    remonCall?.connect(channelId)
-                    remonCall?.onClose {
-                        finish()
-                    }
-                }
-            }
-
-            override fun onPermissionDenied(deniedPermissions: List<String>) {
-                // 접근거부 시 실행할 코드
-                showToast("권한을 허용해주세요!")
-                finish()
-            }
-        }
-
-        TedPermission.with(this)
-            .setPermissionListener(permissionListener)
-            .setDeniedMessage("접근 거부하셨습니다.\n[설정] - [권한]에서 권한을 허용해주세요.")
-            .setPermissions(
-                Manifest.permission.CAMERA,
-                Manifest.permission.RECORD_AUDIO,
-                Manifest.permission.BLUETOOTH
-            )
-            .check()
 
         with(binding) {
             backBtn.setOnClickListener {
@@ -79,6 +40,24 @@ class CallActivity(override val layoutResId: Int = R.layout.activity_call) :
                 showToast("통화가 종료되었어요.")
                 remonCall?.close()
             }
+
+                remonCall = RemonCall.builder()
+                    .context(this@CallActivity)
+                    .localView(localView)        //나의 Video Renderer
+                    .remoteView(remoteView)      //상대방 video Renderer
+                    .videoCodec("VP8")
+                    .videoWidth(640)
+                    .videoHeight(480)
+                    .serviceId("SERVICEID1")    // RemoteMonster 사이트에서 등록했던 당신의 id를 입력하세요.
+                    .key("1234567890")    // RemoteMonster로부터 받은 당신의 key를 입력하세요.
+                    .build()
+                val channelId = "unithon"//intent.getStringExtra("channelId")
+
+                remonCall?.connect(channelId)
+                remonCall?.onClose {
+                    finish()
+                }
+
         }
 
     }
