+++ app/src/main/res/color/color_tab_selected.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="#007AFF" android:state_selected="true"/>
+    <item android:color="#AEAEB2"/>
+</selector>
