+++ app/src/main/res/drawable/ic_friends.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="28dp"
+    android:height="18dp"
+    android:viewportWidth="28"
+    android:viewportHeight="18">
+  <path
+      android:pathData="M19.0977,8.8516C21.1875,8.8516 22.9941,6.9766 22.9941,4.5352C22.9941,2.123 21.1777,0.3359 19.0977,0.3359C17.0078,0.3359 15.1816,2.1621 15.1914,4.5547C15.1914,6.9766 16.998,8.8516 19.0977,8.8516ZM7.8477,9.0762C9.6738,9.0762 11.2461,7.4356 11.2461,5.3164C11.2461,3.2168 9.6641,1.6641 7.8477,1.6641C6.041,1.6641 4.4394,3.2559 4.4492,5.3359C4.4492,7.4356 6.0313,9.0762 7.8477,9.0762ZM2.4082,17.8945H9.8203C8.8047,16.4199 10.0449,13.4512 12.1543,11.8301C11.0605,11.1074 9.6641,10.5703 7.8477,10.5703C3.4434,10.5703 0.6895,13.8223 0.6895,16.5273C0.6895,17.4062 1.168,17.8945 2.4082,17.8945ZM12.9551,17.8945H25.2207C26.7637,17.8945 27.3105,17.4551 27.3105,16.5957C27.3105,14.0762 24.1562,10.5996 19.0879,10.5996C14.0293,10.5996 10.875,14.0762 10.875,16.5957C10.875,17.4551 11.4219,17.8945 12.9551,17.8945Z"
+      android:fillColor="#AEAEB2"/>
+</vector>
