+++ app/src/main/res/drawable/ic_home.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="25dp"
+    android:height="22dp"
+    android:viewportWidth="25"
+    android:viewportHeight="22">
+  <path
+      android:pathData="M0.8398,10.5137C0.8398,10.9141 1.1523,11.2949 1.6797,11.2949C1.9531,11.2949 2.1777,11.1484 2.3828,10.9824L12.1582,2.7793C12.373,2.584 12.6367,2.584 12.8516,2.7793L22.627,10.9824C22.8223,11.1484 23.0469,11.2949 23.3203,11.2949C23.7793,11.2949 24.1504,11.0117 24.1504,10.543C24.1504,10.25 24.0527,10.0547 23.8477,9.8789L21.1133,7.5742V3.668C21.1133,3.2383 20.8398,2.9648 20.4102,2.9648H19.3652C18.9453,2.9648 18.6621,3.2383 18.6621,3.668V5.5137L13.6816,1.334C12.959,0.7188 12.0605,0.7188 11.3281,1.334L1.1523,9.8789C0.9375,10.0547 0.8398,10.2891 0.8398,10.5137ZM3.8867,19.2246C3.8867,20.582 4.707,21.3828 6.084,21.3828H18.9258C20.3027,21.3828 21.1133,20.582 21.1133,19.2246V11.4902L13.0859,4.7617C12.7148,4.4492 12.2754,4.459 11.9141,4.7617L3.8867,11.4902V19.2246ZM15.0098,19.8105H9.9707V13.3555C9.9707,12.9062 10.2734,12.6133 10.7227,12.6133H14.2676C14.7168,12.6133 15.0098,12.9062 15.0098,13.3555V19.8105Z"
+      android:fillColor="#007AFF"/>
+</vector>
