+++ app/src/main/res/drawable/ic_settings.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M11.8199,20H8.1799C7.9518,20 7.7306,19.9221 7.5529,19.7792C7.3752,19.6362 7.2517,19.4368 7.2029,19.214L6.7959,17.33C6.2529,17.0921 5.7381,16.7946 5.2609,16.443L3.4239,17.028C3.2064,17.0973 2.9718,17.0902 2.759,17.0078C2.5462,16.9254 2.3679,16.7727 2.2539,16.575L0.4299,13.424C0.317,13.2261 0.2747,12.9958 0.3097,12.7708C0.3448,12.5457 0.4552,12.3392 0.6229,12.185L2.0479,10.885C1.9831,10.2961 1.9831,9.7019 2.0479,9.113L0.6229,7.816C0.455,7.6618 0.3444,7.4551 0.3093,7.2298C0.2743,7.0045 0.3168,6.774 0.4299,6.576L2.2499,3.423C2.3639,3.2253 2.5422,3.0726 2.755,2.9902C2.9678,2.9078 3.2025,2.9007 3.4199,2.97L5.2569,3.555C5.5009,3.375 5.7549,3.207 6.0169,3.055C6.2699,2.913 6.5299,2.784 6.7959,2.669L7.2039,0.787C7.2525,0.5642 7.3757,0.3647 7.5532,0.2215C7.7307,0.0784 7.9518,0.0002 8.1799,0H11.8199C12.0479,0.0002 12.269,0.0784 12.4465,0.2215C12.6241,0.3647 12.7473,0.5642 12.7959,0.787L13.2079,2.67C13.7501,2.9093 14.2648,3.2067 14.7429,3.557L16.5809,2.972C16.7982,2.9029 17.0326,2.9102 17.2452,2.9926C17.4578,3.0749 17.6359,3.2275 17.7499,3.425L19.5699,6.578C19.8019,6.985 19.7219,7.5 19.3769,7.817L17.9519,9.117C18.0167,9.7059 18.0167,10.3001 17.9519,10.889L19.3769,12.189C19.7219,12.507 19.8019,13.021 19.5699,13.428L17.7499,16.581C17.6358,16.7787 17.4576,16.9314 17.2448,17.0138C17.0319,17.0962 16.7973,17.1033 16.5799,17.034L14.7429,16.449C14.266,16.8003 13.7515,17.0975 13.2089,17.335L12.7959,19.214C12.7471,19.4366 12.6238,19.6359 12.4463,19.7788C12.2688,19.9218 12.0478,19.9998 11.8199,20ZM9.9959,6C8.935,6 7.9176,6.4214 7.1675,7.1716C6.4173,7.9217 5.9959,8.9391 5.9959,10C5.9959,11.0609 6.4173,12.0783 7.1675,12.8284C7.9176,13.5786 8.935,14 9.9959,14C11.0568,14 12.0742,13.5786 12.8243,12.8284C13.5745,12.0783 13.9959,11.0609 13.9959,10C13.9959,8.9391 13.5745,7.9217 12.8243,7.1716C12.0742,6.4214 11.0568,6 9.9959,6Z"
+      android:fillColor="#AEAEB2"/>
+</vector>
