+++ app/src/main/res/layout/activity_main.xml
@@ -1,39 +1,66 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
 
     </data>
 
-    <LinearLayout
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/root_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical">
 
-        <fragment
-            android:id="@+id/map_fragment"
+        <androidx.viewpager2.widget.ViewPager2
+            android:id="@+id/view_pager"
             android:layout_width="match_parent"
-            android:layout_height="120dp"
-            android:name="com.naver.maps.map.MapFragment" />
+            android:layout_height="0dp"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toTopOf="@+id/divider"/>
 
-        <!--
-        <androidx.cardview.widget.CardView
-            android:id="@+id/bottom_sheet"
+        <View
+            android:id="@+id/divider"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:descendantFocusability="beforeDescendants"
-            android:focusableInTouchMode="true"
-            android:orientation="vertical"
-            android:padding="10dp"
-            app:behavior_hideable="false"
-            app:behavior_peekHeight="100dp"
-            app:cardCornerRadius="30dp"
-            app:layout_behavior="com.google.android.material.bottomsheet.BottomSheetBehavior">
+            android:layout_height="0.3dp"
+            android:background="#ACACAC"
+            app:layout_constraintBottom_toTopOf="@+id/tab_layout"/>
 
-        </androidx.cardview.widget.CardView>
-        -->
+        <com.google.android.material.tabs.TabLayout
+            android:id="@+id/tab_layout"
+            android:layout_width="match_parent"
+            android:layout_height="?actionBarSize"
+            app:tabIndicatorHeight="0dp"
+            android:background="#F9F9F9"
+            app:tabSelectedTextColor="#007AFF"
+            app:tabIconTint="@color/color_tab_selected"
+            app:tabTextColor="#AEAEB2"
+            app:tabTextAppearance="@style/tabTextSizeBold"
+            app:layout_constraintBottom_toBottomOf="parent">
 
-    </LinearLayout>
+            <com.google.android.material.tabs.TabItem
+                android:id="@+id/home_tab"
+                android:layout_width="0dp"
+                android:icon="@drawable/ic_home"
+                android:text="Home"
+                android:layout_height="match_parent"/>
+
+            <com.google.android.material.tabs.TabItem
+                android:id="@+id/friends_tab"
+                android:layout_width="0dp"
+                android:icon="@drawable/ic_friends"
+                android:text="Friends"
+                android:layout_height="match_parent"/>
+
+            <com.google.android.material.tabs.TabItem
+                android:id="@+id/setting_tab"
+                android:layout_width="0dp"
+                android:icon="@drawable/ic_settings"
+                android:text="Settings"
+                android:textSize="10sp"
+                android:layout_height="match_parent"/>
+
+        </com.google.android.material.tabs.TabLayout>
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
